+++ lib/control/enum/debug_command.dart
@@ -1,3 +1,3 @@
 enum DebugCommand {
-  none, reload, reset, heal, sethp, setstam, map, setstr;
+  none, reload, reset, heal, sethp, setstam, map, setstr, save;
 }
